Modificara modulos para mejorar UX
- Se ha mejorado la experiencia de usuario al incluir bloques try/catch en los modales
- Se han borrado console.log's
