+++ src/modules/admin/pages/client/client-page.jsx
@@ -1,54 +1,57 @@
-import { useEffect, useState } from "react";
-import { useServiceTypeStore } from "../../../../hooks";
-import { useScreenSizes } from "../../../../shared/constants/screen-width";
-import { Box, Button, CircularProgress, TextField, Typography } from "@mui/material";
-import { AddCircleOutline, Edit } from "@mui/icons-material";
-import { TableComponent } from "../../../../shared/ui/components";
-import { ServiceTypeModal } from "../../components";
+import { useEffect, useState } from 'react';
+import { Box, Typography, Button, TextField, CircularProgress } from '@mui/material';
+import { AddCircleOutline, Edit } from '@mui/icons-material';
+import { useClientStore, useUsersStore } from '../../../../hooks';
+import { TableComponent } from '../../../../shared/ui/components';
+import { useScreenSizes } from '../../../../shared/constants/screen-width';
+import { ClientModal } from '../../components/client/client-modal';
 
-
-export const ServiceTypePage = () => {
+export const ClientPage = () => {
   const {
-    serviceTypes, 
+    clients,
     total,
     loading,
     searchTerm,
     rowsPerPage,
-    currentPage,
+    currentPage, 
     orderBy,
     order,
-    selected,
+    selected, 
     setSearchTerm,
     setRowsPerPageGlobal,
     setPageGlobal,
     setOrderBy,
     setOrder,
-    startLoadingServiceTypePaginated,
-    setSelectedServiceType,
-  } = useServiceTypeStore();
+    startLoadingClientsPaginated,
+    setSelectedClient,
+  } = useClientStore();
   const { isLg } = useScreenSizes();
-  
-  const [isModalOpen, setIsModalOpen] = useState(false);
+
+  const [isModalOpen, setIsModalOpen] = useState(false); 
 
   useEffect(() => {
-    startLoadingServiceTypePaginated();
+    startLoadingClientsPaginated();
   }, [currentPage, rowsPerPage, searchTerm, orderBy, order]);
 
   const openModal = (payload) => {
-    setSelectedServiceType(payload);
-    setIsModalOpen(true);
+    setSelectedClient(payload);
+    setIsModalOpen(true); 
   };
 
   const columns = [
-    { id: "name", label: "Nombre", sortable: true },
-    { id: "description", label: "Descripción", sortable: true },
-    { id: "status", label: "Estado", sortable: true },
+    { id: 'first_name', label: 'Nombre', sortable: true, width: '140px', truncate: true },
+    { id: 'last_name', label: 'Apellido', sortable: true, width: '140px', truncate: true },
+    { id: 'email', label: 'Correo', sortable: true, width: '140px', truncate: true },
+    { id: 'phone', label: '# Telefono', sortable: true, width: '140px', truncate: true },
+    { id: 'document_type', label: 'Tipo Doc', sortable: true, width: '120px', truncate: true },
+    { id: 'document_number', label: '# Documento', sortable: true, width: '140px', truncate: true },
+    { id: 'status', label: 'Estado', sortable: true, width: '140px', truncate: true },
   ];
 
   const actions = [
-    {
-      label: "Editar",
-      icon: <Edit />,
+    { 
+      label: 'Editar', 
+      icon: <Edit />, 
       onClick: (row) => openModal(row),
     },
   ];
@@ -63,8 +66,8 @@ export const ServiceTypePage = () => {
       >
         <Box display="flex" justifyContent="space-between" alignItems="center" sx={{ px: 3, py: 2 }}>
           <Box>
-            <Typography sx={{ fontWeight: 600, fontSize: 24 }}>Listado de Tipos de Servicios</Typography>
-            <Typography sx={{ color: 'text.secondary', fontSize: 16 }}>Administra los tipos de servicios</Typography>
+            <Typography sx={{ fontWeight: 600, fontSize: 24 }}>Listado de Clientes</Typography>
+            <Typography sx={{ color: 'text.secondary', fontSize: 16 }}>Administra los clientes</Typography>
           </Box>
           <Button
             variant="contained"
@@ -72,7 +75,7 @@ export const ServiceTypePage = () => {
             sx={{ backgroundColor: '#212121', color: '#fff', borderRadius: 2, textTransform: 'none', px: 3, py: 1.5 }}
             onClick={() => openModal()} 
           >
-            {isLg ? 'Agregar Tipo de Servicio' : 'Agregar'}
+            {isLg ? 'Agregar Cliente' : 'Agregar'}
           </Button>
         </Box>
 
@@ -95,7 +98,7 @@ export const ServiceTypePage = () => {
           <Box display="flex" justifyContent="center" alignItems="center" sx={{ py: 5 }}>
             <CircularProgress />
           </Box>
-        ) : serviceTypes.length === 0 ? (
+        ) : clients.length === 0 ? (
           <Box display="flex" justifyContent="center" alignItems="center" sx={{ py: 5 }}>
             <Typography sx={{ color: 'text.secondary', fontSize: 16 }}>
               No se encontraron resultados.
@@ -103,7 +106,7 @@ export const ServiceTypePage = () => {
           </Box>
         ) : (
           <TableComponent
-            rows={serviceTypes}
+            rows={clients}
             columns={columns}
             order={order}
             orderBy={orderBy}
@@ -126,13 +129,14 @@ export const ServiceTypePage = () => {
         )}
       </Box>
 
-      <ServiceTypeModal
+      <ClientModal
         open={isModalOpen}
         onClose={() => setIsModalOpen(false)}
-        serviceType={selected}
-        setServiceType={setSelectedServiceType}
+        client={selected}
+        setClient={setSelectedClient}
+        loading={loading}
       />
+
     </Box>
   );
 };
-
