+++ src/modules/admin/pages/client/index.js
@@ -1,0 +1,1 @@
+export * from './client-page';
