+++ src/modules/admin/pages/dashboard/dashboard-page.jsx
@@ -2,6 +2,6 @@ import React from 'react'
 
 export const DashboardPage = () => {
   return (
-      <p>Dashboard Page</p>
+      <p>Dashboard Page :D</p>
   )
 }
