+++ src/modules/admin/pages/dashboard/index.js
@@ -1,0 +1,1 @@
+export * from './dashboard-page';
