+++ src/modules/admin/pages/index.js
@@ -1,7 +1,7 @@
-export * from './dashboard-page';
+export * from './client';
+export * from './dashboard';
 export * from './events';
 export * from './workers';
-export * from './client-page'
 export * from './resource';
-export * from './storehouse-page';
+export * from './storehouse';
 export * from './service';
