+++ src/modules/admin/pages/service/index.js
@@ -1,5 +1,3 @@
-export * from './provider-page';
-export * from './service-add-page';
-export * from './service-edit-page';
-export * from './service-page';
-export * from './service-type-page';
+export * from './provider';
+export * from './service';
+export * from './service-type';
