+++ src/modules/admin/pages/service/provider/index.js
@@ -1,0 +1,1 @@
+export * from './provider-page';
