+++ src/modules/admin/pages/service/service-type/index.js
@@ -1,0 +1,1 @@
+export * from './service-type-page';
