+++ src/modules/admin/pages/service/service/index.js
@@ -1,0 +1,4 @@
+export * from './service-add-page';
+export * from './service-detail-page';
+export * from './service-edit-page';
+export * from './service-page';
