+++ src/modules/admin/pages/service/service/service-add-page.jsx
@@ -1,5 +1,5 @@
 import { Box, Typography, TextField, Button } from '@mui/material';
-import { useServiceStore } from '../../../../hooks';
+import { useServiceStore } from '../../../../../hooks';
 
 export const ServiceAddPage = () => {
   const { startCreateService } = useServiceStore();
