+++ src/modules/admin/pages/service/service/service-detail-page.jsx
