+++ src/modules/admin/pages/service/service/service-page.jsx
@@ -1,14 +1,14 @@
 import { useEffect, useState } from 'react';
 import { Box, Typography, Button, TextField, CircularProgress } from '@mui/material';
 import { AddCircleOutline, Edit } from '@mui/icons-material';
-import { useClientStore, useUsersStore } from '../../../hooks';
-import { TableComponent } from '../../../shared/ui/components';
-import { useScreenSizes } from '../../../shared/constants/screen-width';
-import { ClientModal } from '../components/client/client-modal';
+import { useServiceStore } from '../../../../../hooks';
+import { TableComponent } from '../../../../../shared/ui/components';
+import { useScreenSizes } from '../../../../../shared/constants/screen-width';
+import { Link } from 'react-router-dom';
 
-export const ClientPage = () => {
+export const ServicePage = () => {
   const {
-    clients,
+    services,
     total,
     loading,
     searchTerm,
@@ -22,37 +22,26 @@ export const ClientPage = () => {
     setPageGlobal,
     setOrderBy,
     setOrder,
-    startLoadingClientsPaginated,
-    setSelectedClient,
-  } = useClientStore();
+    startLoadingServicePaginated,
+    setSelectedService,
+  } = useServiceStore();
   const { isLg } = useScreenSizes();
 
-  const [isModalOpen, setIsModalOpen] = useState(false); 
-
   useEffect(() => {
-    startLoadingClientsPaginated();
+    startLoadingServicePaginated();
   }, [currentPage, rowsPerPage, searchTerm, orderBy, order]);
 
-  const openModal = (payload) => {
-    setSelectedClient(payload);
-    setIsModalOpen(true); 
-  };
-
   const columns = [
-    { id: 'first_name', label: 'Nombre', sortable: true, width: '140px', truncate: true },
-    { id: 'last_name', label: 'Apellido', sortable: true, width: '140px', truncate: true },
-    { id: 'email', label: 'Correo', sortable: true, width: '140px', truncate: true },
-    { id: 'phone', label: '# Telefono', sortable: true, width: '140px', truncate: true },
-    { id: 'document_type', label: 'Tipo Doc', sortable: true, width: '120px', truncate: true },
-    { id: 'document_number', label: '# Documento', sortable: true, width: '140px', truncate: true },
-    { id: 'status', label: 'Estado', sortable: true, width: '140px', truncate: true },
+    { id: 'provider_name', label: 'Proveedor', sortable: true },
+    { id: 'service_type_name', label: 'Tipo de Servicio', sortable: true },
+    { id: 'status', label: 'Estado' , sortable: true },
   ];
 
   const actions = [
     { 
       label: 'Editar', 
       icon: <Edit />, 
-      onClick: (row) => openModal(row),
+      url: (row) => `/admin/service/${row._id}`,
     },
   ];
 
@@ -66,17 +55,18 @@ export const ClientPage = () => {
       >
         <Box display="flex" justifyContent="space-between" alignItems="center" sx={{ px: 3, py: 2 }}>
           <Box>
-            <Typography sx={{ fontWeight: 600, fontSize: 24 }}>Listado de Clientes</Typography>
-            <Typography sx={{ color: 'text.secondary', fontSize: 16 }}>Administra los clientes</Typography>
+            <Typography sx={{ fontWeight: 600, fontSize: 24 }}>Listado de Servicios</Typography>
+            <Typography sx={{ color: 'text.secondary', fontSize: 16 }}>Administra los servicios</Typography>
           </Box>
-          <Button
-            variant="contained"
-            startIcon={<AddCircleOutline />}
-            sx={{ backgroundColor: '#212121', color: '#fff', borderRadius: 2, textTransform: 'none', px: 3, py: 1.5 }}
-            onClick={() => openModal()} 
-          >
-            {isLg ? 'Agregar Cliente' : 'Agregar'}
-          </Button>
+          <Link to="/admin/service/new" style={{ textDecoration: 'none', color: 'inherit' }}>
+            <Button
+              variant="contained"
+              startIcon={<AddCircleOutline />}
+              sx={{ backgroundColor: '#212121', color: '#fff', borderRadius: 2, textTransform: 'none', px: 3, py: 1.5 }}
+            >
+              {isLg ? 'Agregar Servicio' : 'Agregar'}
+            </Button>
+          </Link>
         </Box>
 
         <Box
@@ -98,7 +88,7 @@ export const ClientPage = () => {
           <Box display="flex" justifyContent="center" alignItems="center" sx={{ py: 5 }}>
             <CircularProgress />
           </Box>
-        ) : clients.length === 0 ? (
+        ) : services.length === 0 ? (
           <Box display="flex" justifyContent="center" alignItems="center" sx={{ py: 5 }}>
             <Typography sx={{ color: 'text.secondary', fontSize: 16 }}>
               No se encontraron resultados.
@@ -106,7 +96,7 @@ export const ClientPage = () => {
           </Box>
         ) : (
           <TableComponent
-            rows={clients}
+            rows={services}
             columns={columns}
             order={order}
             orderBy={orderBy}
@@ -115,28 +105,19 @@ export const ClientPage = () => {
               setOrder(isAsc ? 'desc' : 'asc');
               setOrderBy(prop);
             }}
-            page={currentPage} 
+            page={currentPage}
             rowsPerPage={rowsPerPage}
             total={total}
             onPageChange={(_, newPage) => setPageGlobal(newPage)}
             onRowsPerPageChange={(e) => {
-              setRowsPerPageGlobal(parseInt(e.target.value, 10)); 
-              setPageGlobal(0); 
+              setRowsPerPageGlobal(parseInt(e.target.value, 10));
+              setPageGlobal(0);
             }}
             actions={actions}
             hasActions
           />
         )}
       </Box>
-
-      <ClientModal
-        open={isModalOpen}
-        onClose={() => setIsModalOpen(false)}
-        client={selected}
-        setClient={setSelectedClient}
-        loading={loading}
-      />
-
     </Box>
   );
 };
