+++ src/modules/admin/pages/storehouse/index.js
@@ -1,0 +1,1 @@
+export * from './storehouse-page';
